requirements-api.txt
torch==2.7.0
transformers==5.6.1
accelerate==1.13.0
peft==0.19.1
bitsandbytes==0.49.2
fastapi==0.136.1
uvicorn==0.46.0
python-multipart==0.0.26
python-dotenv==1.2.2
huggingface_hub==1.11.0
pydantic==2.13.3
safetensors==0.7.0
tokenizers==0.22.2
tqdm==4.67.3
requests==2.33.1